Calculate time trajectory of solution to a Susceptible-Infectious-Removed (SIR) system by numerically solving the multi-group ordinary differential equations

Calculate time trajectory of solution to a Susceptible-Infectious-Removed (SIR) system by numerically solving the multi-group ordinary differential equations

Usage

simodeSIR(
  transmrates,
  recoveryrate,
  popsize,
  initR,
  initI,
  initV,
  times = NULL
)

Arguments

transmrates

matrix of group-to-group (column-to-row) transmission rates

recoveryrate

inverse of mean infectious period

popsize

the population size of each group

initR

initial number of each group already infected and removed (included in final size)

initI

initial number of each group infectious

initV

initial number of each group vaccinated

times

optional argument for times at which the solution is to be calculated; if NULL (the default) the solution will be computed up to the time at which the total number of infectious individuals is sufficiently close to zero (outbreak extinction)

Value

vector of final sizes (number of infected over whole outbreak) for each group

Examples

simodeSIR(transmrates = matrix(0.2, 2 ,2), recoveryrate = 0.3,
popsize = c(100, 150), initR = c(0, 0), initI = c(0, 1), initV = c(10, 10))
#>           time       S1        S2           I1           I2         R1
#> 1     0.000000 90.00000 139.00000 0.000000e+00 1.000000e+00  0.0000000
#> 2     3.333667 89.55461 138.31212 2.863860e-01 8.101482e-01  0.1590066
#> 3     6.667333 89.00733 137.46688 4.570141e-01 8.411435e-01  0.5356531
#> 4    10.001000 88.34071 136.43732 5.961351e-01 9.704711e-01  1.0631523
#> 5    13.334667 87.53725 135.19643 7.346865e-01 1.152992e+00  1.7280586
#> 6    16.668334 86.58104 133.71960 8.829578e-01 1.370414e+00  2.5360039
#> 7    20.002000 85.45987 131.98803 1.042251e+00 1.612175e+00  3.4978770
#> 8    23.335667 84.16787 129.99260 1.208921e+00 1.868022e+00  4.6232098
#> 9    26.669334 82.70806 127.73800 1.375862e+00 2.125278e+00  5.9160774
#> 10   30.003000 81.09448 125.24592 1.533423e+00 2.368410e+00  7.3720953
#> 11   33.336667 79.35304 122.55635 1.670550e+00 2.580117e+00  8.9764152
#> 12   36.670334 77.52058 119.72623 1.776354e+00 2.743497e+00 10.7030688
#> 13   40.004000 75.64206 116.82497 1.841911e+00 2.844735e+00 12.5160248
#> 14   43.337667 73.76611 113.92766 1.861877e+00 2.875567e+00 14.3720110
#> 15   46.671334 71.93997 111.10729 1.835408e+00 2.834686e+00 16.2246197
#> 16   50.005001 70.20494 108.42763 1.766166e+00 2.727746e+00 18.0288909
#> 17   53.338667 68.59312 105.93826 1.661435e+00 2.565995e+00 19.7454474
#> 18   56.672334 67.12593 103.67227 1.530663e+00 2.364025e+00 21.3434070
#> 19   60.006001 65.81433 101.64657 1.383908e+00 2.137370e+00 22.8017632
#> 20   63.339667 64.66016  99.86403 1.230525e+00 1.900477e+00 24.1093112
#> 21   66.673334 63.65822  98.31659 1.078304e+00 1.665381e+00 25.2634711
#> 22   70.007001 62.79840  96.98865 9.330987e-01 1.441118e+00 26.2684974
#> 23   73.340667 62.06767  95.86007 7.988154e-01 1.233727e+00 27.1335160
#> 24   76.674334 61.45165  94.90866 6.776661e-01 1.046616e+00 27.8706815
#> 25   80.008001 60.93583  94.11200 5.705125e-01 8.811263e-01 28.4936598
#> 26   83.341668 60.50629  93.44861 4.772481e-01 7.370833e-01 29.0164588
#> 27   86.675334 60.15025  92.89871 3.971189e-01 6.133280e-01 29.4526344
#> 28   90.009001 59.85622  92.44461 3.289968e-01 5.081176e-01 29.8147810
#> 29   93.342668 59.61416  92.07076 2.715767e-01 4.194352e-01 30.1142622
#> 30   96.676334 59.41538  91.76375 2.235126e-01 3.452027e-01 30.3611066
#> 31  100.010001 59.25248  91.51217 1.835073e-01 2.834169e-01 30.5640085
#> 32  103.343668 59.11921  91.30634 1.503621e-01 2.322259e-01 30.7304260
#> 33  106.677334 59.01031  91.13815 1.230015e-01 1.899689e-01 30.8666839
#> 34  110.011001 58.92144  91.00089 1.004855e-01 1.551942e-01 30.9780728
#> 35  113.344668 58.84898  90.88898 8.200198e-02 1.266475e-01 31.0690207
#> 36  116.678335 58.78993  90.79779 6.685896e-02 1.032600e-01 31.1432074
#> 37  120.012001 58.74185  90.72353 5.447285e-02 8.413028e-02 31.2036736
#> 38  123.345668 58.70272  90.66309 4.435526e-02 6.850423e-02 31.2529233
#> 39  126.679335 58.67088  90.61392 3.609957e-02 5.575380e-02 31.2930160
#> 40  130.013001 58.64499  90.57393 2.936905e-02 4.535886e-02 31.3256400
#> 41  133.346668 58.62394  90.54141 2.388595e-02 3.689051e-02 31.3521765
#> 42  136.680335 58.60682  90.51498 1.942167e-02 2.999571e-02 31.3737551
#> 43  140.014001 58.59291  90.49350 1.578840e-02 2.438431e-02 31.3912991
#> 44  143.347668 58.58161  90.47604 1.283259e-02 1.981921e-02 31.4055602
#> 45  146.681335 58.57242  90.46185 1.042872e-02 1.610659e-02 31.4171506
#> 46  150.015002 58.56496  90.45032 8.474194e-03 1.308796e-02 31.4265692
#> 47  153.348668 58.55889  90.44096 6.885384e-03 1.063404e-02 31.4342223
#> 48  156.682335 58.55397  90.43335 5.593983e-03 8.639566e-03 31.4404405
#> 49  160.016002 58.54996  90.42717 4.544467e-03 7.018766e-03 31.4454924
#> 50  163.349668 58.54671  90.42214 3.691730e-03 5.701691e-03 31.4495963
#> 51  166.683335 58.54407  90.41807 2.998907e-03 4.631548e-03 31.4529301
#> 52  170.017002 58.54193  90.41475 2.435960e-03 3.762236e-03 31.4556382
#> 53  173.350668 58.54018  90.41206 1.978638e-03 3.056006e-03 31.4578379
#> 54  176.684335 58.53877  90.40988 1.607218e-03 2.482175e-03 31.4596246
#> 55  180.018002 58.53762  90.40810 1.305459e-03 2.016113e-03 31.4610759
#> 56  183.351669 58.53668  90.40666 1.060257e-03 1.637653e-03 31.4622548
#> 57  186.685335 58.53593  90.40549 8.612047e-04 1.330154e-03 31.4632121
#> 58  190.019002 58.53531  90.40454 6.996215e-04 1.080318e-03 31.4639894
#> 59  193.352669 58.53481  90.40376 5.682716e-04 8.775199e-04 31.4646209
#> 60  196.686335 58.53440  90.40314 4.614818e-04 7.128237e-04 31.4651341
#> 61  200.020002 58.53407  90.40263 3.748048e-04 5.789374e-04 31.4655508
#> 62  203.353669 58.53381  90.40221 3.043703e-04 4.701040e-04 31.4658896
#> 63  206.687335 58.53359  90.40188 2.472979e-04 3.819520e-04 31.4661640
#> 64  210.021002 58.53341  90.40160 2.008706e-04 3.102352e-04 31.4663872
#> 65  213.354669 58.53327  90.40138 1.631560e-04 2.519739e-04 31.4665685
#> 66  216.688336 58.53315  90.40120 1.325221e-04 2.046674e-04 31.4667158
#> 67  220.022002 58.53306  90.40105 1.076373e-04 1.662449e-04 31.4668354
#> 68  223.355669 58.53298  90.40094 8.742833e-05 1.350301e-04 31.4669326
#> 69  226.689336 58.53292  90.40084 7.101525e-05 1.096738e-04 31.4670115
#> 70  230.023002 58.53287  90.40076 5.768082e-05 8.908275e-05 31.4670756
#> 71  233.356669 58.53283  90.40070 4.684893e-05 7.235942e-05 31.4671277
#> 72  236.690336 58.53279  90.40065 3.805316e-05 5.877246e-05 31.4671700
#> 73  240.024002 58.53276  90.40060 3.087138e-05 4.767631e-05 31.4672045
#> 74  243.357669 58.53274  90.40057 2.510544e-05 3.877282e-05 31.4672322
#> 75  246.691336 58.53272  90.40054 2.037421e-05 3.146916e-05 31.4672550
#> 76  250.025003 58.53271  90.40052 1.653658e-05 2.554231e-05 31.4672734
#> 77  253.358669 58.53270  90.40050 1.342305e-05 2.073160e-05 31.4672884
#> 78  256.692336 58.53269  90.40049 1.089571e-05 1.682698e-05 31.4673006
#> 79  260.026003 58.53268  90.40047 8.843851e-06 1.365829e-05 31.4673104
#> 80  263.359669 58.53267  90.40046 7.178224e-06 1.108656e-05 31.4673184
#> 81  266.693336 58.53267  90.40046 5.826387e-06 8.998921e-06 31.4673249
#> 82  270.027003 58.53267  90.40045 4.729264e-06 7.304197e-06 31.4673302
#> 83  273.360669 58.53266  90.40044 3.838751e-06 5.928604e-06 31.4673345
#> 84  276.694336 58.53266  90.40044 3.115864e-06 4.812159e-06 31.4673380
#> 85  280.028003 58.53266  90.40044 2.529068e-06 3.906013e-06 31.4673408
#> 86  283.361670 58.53265  90.40043 2.052790e-06 3.170484e-06 31.4673431
#> 87  286.695336 58.53265  90.40043 1.666228e-06 2.573425e-06 31.4673449
#> 88  290.029003 58.53265  90.40043 1.352467e-06 2.088793e-06 31.4673464
#> 89  293.362670 58.53265  90.40043 1.097781e-06 1.695439e-06 31.4673477
#> 90  296.696336 58.53265  90.40043 8.910483e-07 1.376172e-06 31.4673487
#> 91  300.030003 58.53265  90.40043 7.232472e-07 1.117025e-06 31.4673495
#> 92  303.363670 58.53265  90.40043 5.870501e-07 9.066727e-07 31.4673501
#> 93  306.697336 58.53265  90.40042 4.765026e-07 7.359298e-07 31.4673506
#> 94  310.031003 58.53265  90.40042 3.867712e-07 5.973425e-07 31.4673511
#> 95  313.364670 58.53265  90.40042 3.139360e-07 4.848555e-07 31.4673514
#> 96  316.698337 58.53265  90.40042 2.548166e-07 3.935515e-07 31.4673517
#> 97  320.032003 58.53265  90.40042 2.068309e-07 3.194402e-07 31.4673519
#> 98  323.365670 58.53265  90.40042 1.678821e-07 2.592845e-07 31.4673521
#> 99  326.699337 58.53265  90.40042 1.362677e-07 2.104572e-07 31.4673523
#> 100 330.033003 58.53265  90.40042 1.106065e-07 1.708253e-07 31.4673524
#> 101 333.366670 58.53265  90.40042 8.977757e-08 1.386567e-07 31.4673525
#> 102 336.700337 58.53265  90.40042 7.287115e-08 1.125457e-07 31.4673526
#> 103 340.034003 58.53265  90.40042 5.914854e-08 9.135165e-08 31.4673527
#> 104 343.367670 58.53265  90.40042 4.801007e-08 7.414877e-08 31.4673527
#> 105 346.701337 58.53265  90.40042 3.896908e-08 6.018551e-08 31.4673527
#> 106 350.035004 58.53265  90.40042 3.162444e-08 4.884221e-08 31.4673528
#> 107 353.368670 58.53265  90.40042 2.562272e-08 3.957289e-08 31.4673528
#> 108 356.702337 58.53265  90.40042 2.033826e-08 3.141133e-08 31.4673528
#> 109 360.036004 58.53265  90.40042 1.683598e-08 2.600225e-08 31.4673529
#> 110 363.369670 58.53265  90.40042 1.393680e-08 2.152462e-08 31.4673529
#> 111 366.703337 58.53265  90.40042 1.153686e-08 1.781805e-08 31.4673529
#> 112 370.037004 58.53265  90.40042 9.550198e-09 1.474975e-08 31.4673529
#> 113 373.370670 58.53265  90.40042 7.905639e-09 1.220982e-08 31.4673529
#> 114 376.704337 58.53265  90.40042 6.544276e-09 1.010727e-08 31.4673529
#> 115 378.715709 58.53265  90.40042 5.856351e-09 9.044810e-09 31.4673529
#>             R2
#> 1    0.0000000
#> 2    0.8777359
#> 3    1.6919758
#> 4    2.5922061
#> 5    3.6505815
#> 6    4.9099821
#> 7    6.3997995
#> 8    8.1393799
#> 9   10.1367177
#> 10  12.3856682
#> 11  14.8635289
#> 12  17.5302782
#> 13  20.3302986
#> 14  23.1967704
#> 15  26.0580229
#> 16  28.8446201
#> 17  31.4957464
#> 18  33.9637064
#> 19  36.2160563
#> 20  38.2354920
#> 21  40.0180271
#> 22  41.5702355
#> 23  42.9062072
#> 24  44.0447206
#> 25  45.0068733
#> 26  45.8143086
#> 27  46.4879576
#> 28  47.0472726
#> 29  47.5098049
#> 30  47.8910425
#> 31  48.2044130
#> 32  48.4614356
#> 33  48.6718785
#> 34  48.8439124
#> 35  48.9843765
#> 36  49.0989536
#> 37  49.1923403
#> 38  49.2684037
#> 39  49.3303246
#> 40  49.3807107
#> 41  49.4216949
#> 42  49.4550217
#> 43  49.4821175
#> 44  49.5041430
#> 45  49.5220436
#> 46  49.5365902
#> 47  49.5484101
#> 48  49.5580137
#> 49  49.5658159
#> 50  49.5721543
#> 51  49.5773032
#> 52  49.5814856
#> 53  49.5848829
#> 54  49.5876425
#> 55  49.5898840
#> 56  49.5917044
#> 57  49.5931830
#> 58  49.5943838
#> 59  49.5953591
#> 60  49.5961514
#> 61  49.5967951
#> 62  49.5973183
#> 63  49.5977421
#> 64  49.5980869
#> 65  49.5983670
#> 66  49.5985944
#> 67  49.5987792
#> 68  49.5989292
#> 69  49.5990511
#> 70  49.5991501
#> 71  49.5992306
#> 72  49.5992959
#> 73  49.5993492
#> 74  49.5993920
#> 75  49.5994271
#> 76  49.5994556
#> 77  49.5994788
#> 78  49.5994975
#> 79  49.5995128
#> 80  49.5995251
#> 81  49.5995352
#> 82  49.5995433
#> 83  49.5995499
#> 84  49.5995553
#> 85  49.5995596
#> 86  49.5995632
#> 87  49.5995661
#> 88  49.5995684
#> 89  49.5995703
#> 90  49.5995718
#> 91  49.5995731
#> 92  49.5995741
#> 93  49.5995749
#> 94  49.5995756
#> 95  49.5995761
#> 96  49.5995765
#> 97  49.5995769
#> 98  49.5995772
#> 99  49.5995774
#> 100 49.5995776
#> 101 49.5995778
#> 102 49.5995779
#> 103 49.5995780
#> 104 49.5995781
#> 105 49.5995781
#> 106 49.5995782
#> 107 49.5995782
#> 108 49.5995783
#> 109 49.5995783
#> 110 49.5995783
#> 111 49.5995783
#> 112 49.5995784
#> 113 49.5995784
#> 114 49.5995784
#> 115 49.5995784
simodeSIR(transmrates = matrix(0.2, 2 ,2), recoveryrate = 0.3,
popsize = c(100, 150), initR = c(0, 0), initI = c(0, 1), initV = c(10, 10), times = 0:10)
#>    time       S1       S2        I1        I2         R1        R2
#> 1     0 90.00000 139.0000 0.0000000 1.0000000 0.00000000 0.0000000
#> 2     1 89.87609 138.8086 0.1072210 0.9064177 0.01669000 0.2849559
#> 3     2 89.74407 138.6047 0.1936700 0.8479245 0.06226244 0.5473490
#> 4     3 89.60349 138.3876 0.2651112 0.8160189 0.13139526 0.7963634
#> 5     4 89.45391 138.1566 0.3258273 0.8044142 0.22025939 1.0389863
#> 6     5 89.29486 137.9110 0.3789958 0.8084720 0.32614081 1.2805723
#> 7     6 89.12587 137.6500 0.4269846 0.8247553 0.44714292 1.5252861
#> 8     7 88.94646 137.3729 0.4715463 0.8507354 0.58199021 1.7763932
#> 9     8 88.75616 137.0790 0.5139841 0.8845394 0.72985954 2.0365080
#> 10    9 88.55447 136.7675 0.5552629 0.9247796 0.89026573 2.3077591
#> 11   10 88.34093 136.4377 0.5960943 0.9704230 1.06297337 2.5919149